I removed firebas/auth import tag from index.js file. I added REACT_APP_MESSAGING_SENDER_ID='' to the .env file and the value from firebase config. Also reordered import tags in App.js file and renamed the file to App.js from index.js file.
