+++ src/index.js
@@ -2,8 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import firebase from 'firebase/app';
 import firebaseConfig from './helpers/apiKeys';
-import 'firebase/auth';
-import App from './App';
+import App from './App/App';
 import reportWebVitals from './reportWebVitals';
 
 import 'bootstrap/dist/css/bootstrap.min.css';
